requirements.txt
# Core
pandas>=2.0

# GUI
gradio>=4.0

# API
fastapi>=0.100
uvicorn>=0.20

# NLP / Embeddings
sentence-transformers>=2.2,<3
transformers>=4.30,<5
numpy>=1.24

# LLM enrichment (OpenRouter)
httpx>=0.24

# Config / env
python-dotenv>=1.0
pyyaml>=6.0
